+++ anomaly_mle/__init__.py
@@ -1,5 +1,5 @@
 """
-Shape MLE Challenge Solution
+Anomaly detection
 """
 
 __version__ = "0.1.0"
