+++ anomaly_mle/utils/get_session_spark.py
@@ -4,7 +4,7 @@ Module for getting a Spark session
 
 from pyspark.sql import SparkSession
 
-from shape_mle.infrastructure.logging import logger
+from anomaly_mle.infrastructure.logging import logger
 
 def get_spark_session(app_name: str = "ShapeMLE") -> SparkSession:
     """
